- update translations - add tooltips to nav icon buttons - rename `components/global` to `components/layout` - relocated language and light switch buttons
