+++ messages/hi.json
@@ -44,9 +44,14 @@
     "settings": "सेटिंग्स",
     "logout": "लॉगआउट"
   },
-  "language": {
-    "change": "भाषा बदलें",
-    "select": "भाषा चुनें"
+  "language_button": {
+    "tooltip": "भाषा"
+  },
+  "light_switch": {
+    "tooltip": "विषय",
+    "light": "लाइट",
+    "dark": "डार्क",
+    "system": "सिस्टम"
   },
   "signin": {
     "title": "साइन इन",
@@ -151,12 +156,6 @@
       "min_length": "आपका वन-टाइम पासवर्ड कम से कम 6 अक्षर का होना चाहिए",
       "invalid": "आपके द्वारा दर्ज किया गया कोड सत्यापित नहीं हो सका। कृपया फिर से प्रयास करें।"
     }
-  },
-  "theme": {
-    "light_switch": "थीम बदलें",
-    "light": "लाइट",
-    "dark": "डार्क",
-    "system": "सिस्टम"
   },
   "connection": {
     "offline": "आप ऑफ़लाइन हैं",
