+++ src/lib/components/layout/footer.svelte
