+++ src/lib/components/layout/nav-bar/language-button.svelte
@@ -2,6 +2,7 @@
   import { Button } from '$lib/components/ui/button/index.js';
   import { Languages } from 'lucide-svelte';
   import * as DropdownMenu from '$lib/components/ui/dropdown-menu';
+  import * as Tooltip from "@/components/ui/tooltip";
   import { getLocale, setLocale, locales } from '$lib/paraglide/runtime.js';
   import { m } from '$lib/paraglide/messages.js';
 
@@ -10,14 +11,23 @@
 
 <DropdownMenu.Root>
   <DropdownMenu.Trigger>
-    <Button
-      variant="ghost"
-      size="icon"
-      class={`${className} text-muted-foreground hover:text-foreground`}
-      aria-label={m['language.select']()}
-    >
-      <Languages class="h-5 w-5" />
-    </Button>
+    <Tooltip.Provider>
+      <Tooltip.Root>
+        <Tooltip.Trigger>
+          <Button
+            variant="ghost"
+            size="icon"
+            class={`${className} text-muted-foreground hover:text-foreground`}
+            aria-label={m['language_button.tooltip']()}
+          >
+            <Languages class="h-5 w-5" />
+          </Button>
+        </Tooltip.Trigger>
+        <Tooltip.Content>
+          <p>{m['language_button.tooltip']()}</p>
+        </Tooltip.Content>
+      </Tooltip.Root>
+    </Tooltip.Provider>
   </DropdownMenu.Trigger>
   <DropdownMenu.Content>
     {#each locales as locale}
